% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele4_5_6-IonescuCristian.R
\name{centeredMomentCompute}
\alias{centeredMomentCompute}
\title{centeredMomentCompute}
\usage{
centeredMomentCompute(x, ord)
}
\arguments{
\item{x}{Repartition to compute}

\item{ord}{Order moment}
}
\description{
This function compute for a repartition (x) centered moment, for a specified order.
}
\examples{
centeredMomentCompute(rnorm(100), 1)
}
\keyword{VA}
