% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele10_11_12-OrzataAndrei.R
\name{covVAC}
\alias{covVAC}
\title{covVAC}
\usage{
covVAC(f)
}
\arguments{
\item{f}{densitatea comuna a variabilelor continue aleatoare X si Y}
}
\description{
Functia covVAC() preia in f densitatea comuna a variabilelor continue aleatoare X si Y, calculand valorile asteptate / mediile lui X, Y si a combinatiei dintre cele doua  si intoarce conform formulei de mai sus, covariatia.
}
\examples{
covVAC(f)
}
\keyword{VA}
