% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele10_11_12-OrzataAndrei.R
\name{cvVAC}
\alias{cvVAC}
\title{cvVAC}
\usage{
cvVAC(f)
}
\arguments{
\item{f}{densitatea comuna a variabilelor continue aleatoare X si Y}
}
\description{
Functia cvVAC() preia in f densitatea comuna a variabilelor continue aleatoare X si Y, calculeaza deviatia fiecaruia si o stocheaza in variabilele devX, devY si conform formulei de mai sus, utilizand functia de covariatie covVAC(), returneaza valoare coeficientului de corelatie.
}
\examples{
cvVAC(f)
}
\keyword{VA}
