% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele10_11_12-OrzataAndrei.R
\name{dCondXY}
\alias{dCondXY}
\title{dCondXY}
\usage{
dCondXY(f)
}
\arguments{
\item{f}{densitatea comuna a variabilelor continue aleatoare X si Y}
}
\description{
In dCondXY(), dCondYX(), folosind functiile mentionate anterior si media/ valoarea asteptata a densitatii comune a variabilelor continue aleatoare X si Y, se calculeaza conform formulei ai se returneaza valorile densitatii cond in val
}
\examples{
dCondXY(f)
}
\keyword{VA}
