% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele4_5_6-IonescuCristian.R
\name{dispersionCompute}
\alias{dispersionCompute}
\title{dispersionCompute}
\usage{
dispersionCompute(x, g)
}
\arguments{
\item{x}{Repartition to compute}

\item{g}{Repartition function}
}
\description{
Dispersion compute for random variable g(X)
}
\examples{
dispersionCompute(RV, i, j)
}
\keyword{VA}
