% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele10_11_12-OrzataAndrei.R
\name{dMarY}
\alias{dMarY}
\title{dMarY}
\usage{
dMarY(f)
}
\arguments{
\item{f}{densitatea comuna a variabilelor continue aleatoare X si Y}
}
\description{
In functiile dMarY() si dMarX() se preia in f densitatea comuna a variabilelor continue aleatoare X si Y si conform formulei descrise anterior se returneza valorile densitatii marginala in variabila val.
}
\examples{
dMarY(f)
}
\keyword{VA}
