% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele4_5_6-IonescuCristian.R
\name{firstMomentCompute}
\alias{firstMomentCompute}
\title{firstMomentCompute}
\usage{
firstMomentCompute(x, ord)
}
\arguments{
\item{x}{Repartition to compute}

\item{ord}{Order moment}
}
\description{
This function compute for a repartition (x) initial moment, for a specified order.
}
\examples{
firstMomentCompute(rnorm(100), 1)
}
\keyword{VA}
