% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele1_2_3-AnghelacheVlad.R
\name{getNConst}
\alias{getNConst}
\title{getNConst}
\usage{
getNConst(f)
}
\arguments{
\item{f}{a function}
}
\description{
The getNConst function returns a normalizing constant for a positive function. If the function doesn't have one, it returns NULL and prints the following message ‘Nu exista constanta de normalizare pentru functia data!’. If an error occurs, the message for that specific error will also be printed.
}
\examples{
getNConst(f)
}
\keyword{VA}
