% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele1_2_3-AnghelacheVlad.R
\name{isDensity}
\alias{isDensity}
\title{isDensity}
\usage{
isDensity(f)
}
\arguments{
\item{f}{a function.}
}
\description{
This function  check if a function is a probability density function.
}
\examples{
isDensity(f)
}
\keyword{VA}
