% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele1_2_3-AnghelacheVlad.R
\name{makeVA}
\alias{makeVA}
\title{makeVA}
\usage{
makeVA(f)
}
\arguments{
\item{f}{a probability density function.}
}
\description{
The function makeVA  makes a continuous random variable object which contains  the cumulative distribution function r and the probability density function d.
}
\examples{
makeVA(f)
}
\keyword{VA}
