% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele4_5_6-IonescuCristian.R
\name{momentsCompute}
\alias{momentsCompute}
\title{momentsCompute}
\usage{
momentsCompute(x)
}
\arguments{
\item{x}{Repartition to compute}
}
\description{
This function compute for a repartition (x) initial moment and centered moment, from ord 1 to 4.
}
\examples{
momentsCompute(rnorm(100))
}
\keyword{VA}
