% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele4_5_6-IonescuCristian.R
\name{printContRV}
\alias{printContRV}
\title{printContRV}
\usage{
printContRV(RV, i, j)
}
\arguments{
\item{RV}{a continuous random variable}

\item{i}{start of sequence}

\item{j}{end of sequence}
}
\description{
The printContRV function generate the graphic representations of the probability density function and the cumulative distribution function for a given continuous RV
}
\examples{
printContRV(RV, i, j)
}
\keyword{VA}
