% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele4_5_6-IonescuCristian.R
\name{printRandomDp}
\alias{printRandomDp}
\title{printRandomDp}
\usage{
printRandomDp()
}
\description{
This function generate some prints on density and distribution function for multiple random repartitions.
}
\examples{
printRandomDp()
}
\keyword{VA}
