% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerintele10_11_12-OrzataAndrei.R
\name{SumDif}
\alias{SumDif}
\title{SumDif}
\usage{
SumDif(funcX, funcY)
}
\arguments{
\item{funcX}{densitatea unei variabile continue aleatoare x}

\item{funcY}{densitatea unei variabile continue aleatoare Y}
}
\description{
Functia SumDif() preia densitatile a doua variabile aleatoare continue X si Y, in funcX() si funcY(), si  folosind formulele matematice anterioare calculeaza densitatea sumei acestora in funcZT(), precum variatia si media diferentelor sumei si diferente, urmand ca acestea sa fie salvate in variabile miux, miuy si varx, vary respectiv, urmand ca acestea sa fie calculate pentru suma si diferenta si afiaate in consola cu ajutorul functiei print().
}
\examples{
SumDif(f)
}
\keyword{VA}
